fix undo, getpiecemap, piece row col after move, add 1.5 multiplier for doubles

+++ ai.h
@@ -14,5 +14,5 @@ public:
 private:
     Move randomMove(const Board &board) const;
     void orderMoves();
-    int negamax(Board board, int depth, int color);
+    float negamax(Board board, int depth, int color);
 };
